+++ src/app/page.tsx
@@ -1,15 +1,15 @@
-import AboutTracer from "@/components/RouteHome/aboutTracer";
-import CoreChallengeSection from "@/components/RouteHome/core-challenges";
-import DataPrivacyNotice from "@/components/RouteHome/data-privacy-notice";
-import DigitalDiscoverySection from "@/components/RouteHome/digitalDiscoverySection";
-import Faq from "@/components/RouteHome/faq";
-import HeroSection from "@/components/RouteHome/hero";
-import HowTracerWorks from "@/components/RouteHome/howTracerWorks";
-import InfrastructureReview from "@/components/RouteHome/infrastructure-review";
-import ModernMonitoringPlatform from "@/components/RouteHome/modernMonitoringPlatform";
-import PersonalizedDemoHome from "@/components/RouteHome/personalized-demo-home";
-import UnifiedPlatformOverview from "@/components/RouteHome/unified-platform-overview";
-import WhyMonitoringSection from "@/components/RouteHome/whyMonitoringSection";
+import AboutTracer from "@/components/RouteHome/AboutTracer";
+import CoreChallengeSection from "@/components/RouteHome/CoreChallenges";
+import DataPrivacyNotice from "@/components/RouteHome/DataPrivacyNotice";
+import DigitalDiscoverySection from "@/components/RouteHome/DigitalDiscoverySection";
+import Faq from "@/components/RouteHome/Faq";
+import HeroSection from "@/components/RouteHome/Hero";
+import HowTracerWorks from "@/components/RouteHome/HowTracerWorks";
+import InfrastructureReview from "@/components/RouteHome/InfrastructureReview";
+import ModernMonitoringPlatform from "@/components/RouteHome/ModernMonitoringPlatform";
+import PersonalizedDemoHome from "@/components/RouteHome/PersonalizedDemoHome";
+import UnifiedPlatformOverview from "@/components/RouteHome/UnifiedPlatformOverview";
+import WhyMonitoringSection from "@/components/RouteHome/WhyMonitoringSection";
 
 // Shared
 import PreLoaderProgress from "@/components/shared/PreLoader"
