+++ src/components/MdxProvider.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { MDXProvider } from '@mdx-js/react';
-import { useMDXComponents } from './mdx-components';
+import { useMDXComponents } from './MdxComponents';
 
 interface MDXProviderWrapperProps {
   children: React.ReactNode;
@@ -10,7 +10,7 @@ interface MDXProviderWrapperProps {
 
 export function MDXProviderWrapper({ children }: MDXProviderWrapperProps) {
   const components = useMDXComponents({});
-  
+
   return (
     <MDXProvider components={components}>
       {children}
