+++ src/components/RouteHome/CoreChallenges.tsx
