+++ src/components/RouteHome/ModernMonitoringPlatform.tsx
