+++ src/components/RoutePlatform/Approach.tsx
