+++ src/components/RoutePlatform/GainVisibility.tsx
