+++ src/components/RoutePlatform/Hero.tsx
