+++ src/components/RoutePlatform/Monitoring.tsx
