+++ src/components/RoutePlatform/YourPartner.tsx
