+++ src/components/careers/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { motion } from "framer-motion";
 import Link from "next/link";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/Button";
 
 const CareersPage: React.FC = () => {
   const fadeInUp = {
