+++ src/components/shared/Navbar.tsx
@@ -6,7 +6,7 @@ import {
     SheetHeader,
     SheetTitle,
     SheetTrigger
-} from "@/components/ui/sheet"
+} from "@/components/ui/Sheet"
 import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
